consolidate more places where metamodel is used
