+++ tests/test_base/test_metamodel.py
@@ -1,23 +1,30 @@
 import unittest
 from typing import List
+import pytest
 
 from pyshex.shex_evaluator import EvaluationResult
 
-from linkml import LOCAL_METAMODEL_LDCONTEXT_FILE, LOCAL_METAMODEL_YAML_FILE
+from linkml import LOCAL_METAMODEL_LDCONTEXT_FILE, LOCAL_METAMODEL_YAML_FILE, METAMODEL_NAMESPACE
 from linkml.generators.markdowngen import MarkdownGenerator
 from linkml.generators.owlgen import OwlSchemaGenerator
 from linkml.generators.rdfgen import RDFGenerator
 from linkml.generators.shexgen import ShExGenerator
+from linkml.generators.jsonldcontextgen import ContextGenerator
+from linkml.generators.pythongen import PythonGenerator
 from tests import SKIP_MARKDOWN_VALIDATION, SKIP_MARKDOWN_VALIDATION_REASON
 from tests.test_base.environment import env
 from tests.utils.compare_rdf import compare_rdf
 from tests.utils.generatortestcase import GeneratorTestCase
+from tests.utils.filters import ldcontext_metadata_filter
+from tests.utils.filters import metadata_filter
+from tests.utils.python_comparator import compare_python
 
 
 class MetaModelTestCase(GeneratorTestCase):
     env = env
     model_name = "meta"
 
+    @pytest.mark.slow
     @unittest.skipIf(SKIP_MARKDOWN_VALIDATION, SKIP_MARKDOWN_VALIDATION_REASON)
     def test_meta_markdown(self):
         """Test the markdown generator for the biolink model"""
@@ -28,6 +35,7 @@ class MetaModelTestCase(GeneratorTestCase):
             input_file=LOCAL_METAMODEL_YAML_FILE,
         )
 
+    @pytest.mark.slow
     def test_meta_owl_schema(self):
         """Test the owl schema generator for the linkml model"""
         self.single_file_generator(
@@ -51,14 +59,17 @@ class MetaModelTestCase(GeneratorTestCase):
                     print(r.reason)
         return success
 
+    @pytest.mark.slow
     def test_meta_shexc(self):
         """Test the shex ShExC generation"""
         self.single_file_generator("shex", ShExGenerator, format="shex", yaml_file=LOCAL_METAMODEL_YAML_FILE)
 
+    @pytest.mark.slow
     def test_meta_shecj(self):
         """Test the shex ShExJ generation"""
         self.single_file_generator("shexj", ShExGenerator, format="json", yaml_file=LOCAL_METAMODEL_YAML_FILE)
 
+    @pytest.mark.slow
     def test_meta_rdf(self):
         """Test the rdf generator for the biolink model"""
 
@@ -83,6 +94,30 @@ class MetaModelTestCase(GeneratorTestCase):
         #     self.assertTrue(self._evaluate_shex_results(results))
         # else:
         #     print("*** RDF Model validation step was skipped. Set: tests.__init__.DO_SHEX_VALIDATION to run it")
+
+    @pytest.mark.slow
+    def test_metamodel_context(self):
+        """Build meta.context.jsonld"""
+        self.model_name = "meta"
+        self.single_file_generator(
+            "context.jsonld",
+            ContextGenerator,
+            yaml_file=LOCAL_METAMODEL_YAML_FILE,
+            serialize_args=dict(base=METAMODEL_NAMESPACE),
+            filtr=ldcontext_metadata_filter,
+        )
+
+    @pytest.mark.slow
+    def test_metamodel_python(self):
+        """Build meta.py"""
+        self.env.generate_single_file(
+            "meta.py",
+            lambda: PythonGenerator(env.meta_yaml, importmap=env.import_map, genmeta=True).serialize(),
+            value_is_returned=True,
+            filtr=metadata_filter,
+            comparator=lambda exp, act: compare_python(exp, act, self.env.expected_path("meta.py")),
+            use_testing_root=True,
+        )
 
 
 if __name__ == "__main__":
